project-1-e-invoicing/requirements.txt
# Core dependencies
Flask==2.2.5
functions-framework==3.5.0
Jinja2==3.1.3
pandas==2.2.2
requests==2.31.0

# Google API dependencies
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
google-cloud-secret-manager==2.23.3